型変更

+++ Assets/Scripts/PrefectureData.cs
@@ -5,8 +5,8 @@ public class PrefectureData
 {
     public readonly int id;
     public readonly string name;
-    public readonly float longitude;
-    public readonly float latitude;
+    public readonly double longitude;
+    public readonly double latitude;
 
     public GameObject ownObj;
 
@@ -14,7 +14,7 @@ public class PrefectureData
     {
         id = int.Parse(csvLine[0]);
         name = csvLine[1];
-        latitude = float.Parse(csvLine[2]);
-        longitude = float.Parse(csvLine[3]);
+        latitude = double.Parse(csvLine[2]);
+        longitude = double.Parse(csvLine[3]);
     }
 }
